use-cases/04-edge-ai-inference/main.py
"""Edge AI Inference Platform - On-device AI without cloud dependency.

Demonstrates production edge AI patterns:
- Model optimization (static shape, quantization)
- Hardware-agnostic inference pipeline
- Real-time detection with NMS
- Performance benchmarking

API:
    POST /api/v1/detect          - Run inference on uploaded image
    POST /api/v1/detect/random   - Run inference on synthetic data
    GET  /api/v1/model/info      - Model information
    POST /api/v1/model/benchmark - Run performance benchmark
    POST /api/v1/model/optimize  - Simulate model optimization
    GET  /api/v1/health          - Health check
    GET  /                       - Interactive demo
"""

import io
import logging
import time

import numpy as np
from fastapi import FastAPI, File, UploadFile
from fastapi.responses import HTMLResponse

from config import config
from inference.engine import EdgeInferenceEngine
from optimizer.model_optimizer import optimize_model, OptimizationConfig, get_optimization_recommendations
from models.schemas import BenchmarkResult, HealthResponse, InferenceResult, ModelInfo

logging.basicConfig(level=getattr(logging, config.LOG_LEVEL), format="%(asctime)s [%(name)s] %(levelname)s: %(message)s")
logger = logging.getLogger(__name__)

app = FastAPI(title="Edge AI Inference Platform", version="1.0.0")
engine: EdgeInferenceEngine | None = None


@app.on_event("startup")
async def startup():
    global engine
    engine = EdgeInferenceEngine(backend=config.INFERENCE_BACKEND)
    logger.info(f"Edge AI Platform started | Backend: {config.INFERENCE_BACKEND} | Port: {config.PORT}")


@app.post("/api/v1/detect", response_model=InferenceResult)
async def detect_image(file: UploadFile = File(...)):
    """Run object detection on uploaded image."""
    content = await file.read()
    try:
        from PIL import Image
        img = Image.open(io.BytesIO(content)).convert("RGB")
        image_data = np.array(img)
    except Exception:
        # Fallback: treat as raw numpy
        image_data = np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8)
    return engine.detect(image_data)


@app.post("/api/v1/detect/random", response_model=InferenceResult)
async def detect_random(width: int = 640, height: int = 480):
    """Run inference on synthetic random image (for testing)."""
    image_data = np.random.randint(0, 255, (height, width, 3), dtype=np.uint8)
    return engine.detect(image_data)


@app.get("/api/v1/model/info", response_model=ModelInfo)
async def model_info():
    return ModelInfo(
        name="edge-detector-v1",
        backend=engine.backend,
        input_shape=list(engine.input_shape),
        num_classes=engine.num_classes,
        status="loaded",
        optimizations=["static_shape", "numpy_demo"],
    )


@app.post("/api/v1/model/benchmark", response_model=BenchmarkResult)
async def run_benchmark(iterations: int = 50):
    """Run inference benchmark."""
    result = engine.benchmark(iterations=min(iterations, 500))
    return BenchmarkResult(**result)


@app.post("/api/v1/model/optimize")
async def optimize(quantization: str = "fp16", static_shape: bool = True):
    """Simulate model optimization and show expected improvements."""
    opt_config = OptimizationConfig(
        quantization=quantization,
        static_shape=static_shape,
        backend=config.INFERENCE_BACKEND,
    )
    result = optimize_model("edge-detector-v1", opt_config)
    recommendations = get_optimization_recommendations(
        inference_ms=10, target_ms=5
    )
    return {
        "optimization_result": {
            "original_size_mb": result.original_size_mb,
            "optimized_size_mb": result.optimized_size_mb,
            "size_reduction_pct": result.size_reduction_pct,
            "estimated_speedup": f"{result.estimated_speedup}x",
            "optimizations_applied": result.optimizations_applied,
        },
        "recommendations": recommendations,
    }


@app.get("/api/v1/health", response_model=HealthResponse)
async def health():
    return HealthResponse(status="ok", backend=engine.backend, models_loaded=1)


@app.get("/", response_class=HTMLResponse)
async def demo():
    return """<!DOCTYPE html>
<html><head><meta charset="UTF-8"><title>Edge AI Inference</title>
<style>
*{margin:0;padding:0;box-sizing:border-box}
body{font-family:-apple-system,sans-serif;background:#0f172a;color:#e2e8f0;min-height:100vh}
.c{max-width:900px;margin:0 auto;padding:2rem}
h1{font-size:2rem;color:#f97316;margin-bottom:.5rem}
.sub{color:#94a3b8;margin-bottom:2rem}
.grid{display:grid;grid-template-columns:repeat(3,1fr);gap:1rem;margin-bottom:1.5rem}
.card{background:#1e293b;border:1px solid #334155;border-radius:8px;padding:1rem;text-align:center}
.card .val{font-size:1.5rem;font-weight:700;color:#f97316}.card .lbl{font-size:.8rem;color:#94a3b8}
.section{background:#1e293b;border:1px solid #334155;border-radius:8px;padding:1.5rem;margin-bottom:1.5rem}
h2{color:#f97316;margin-bottom:1rem;font-size:1.1rem}
button{padding:.5rem 1.5rem;border:none;border-radius:8px;background:#ea580c;color:white;
       font-size:.9rem;cursor:pointer;font-weight:600;margin:.5rem .5rem 0 0}
button:hover{background:#c2410c}
#output{background:#0f172a;border:1px solid #334155;border-radius:8px;padding:1rem;
        min-height:150px;white-space:pre-wrap;font-family:monospace;font-size:.85rem;margin-top:1rem}
select{padding:.5rem;border:1px solid #334155;border-radius:8px;background:#1e293b;color:#e2e8f0;margin-right:.5rem}
</style></head>
<body><div class="c">
<h1>Edge AI Inference Platform</h1>
<p class="sub">On-device AI inference without cloud dependency | YOLO-style object detection</p>

<div class="grid" id="stats">
<div class="card"><div class="val" id="backend">numpy</div><div class="lbl">Backend</div></div>
<div class="card"><div class="val" id="classes">10</div><div class="lbl">Classes</div></div>
<div class="card"><div class="val" id="shape">640x640</div><div class="lbl">Input Shape</div></div>
</div>

<div class="section">
<h2>Inference</h2>
<button onclick="runDetection()">Detect (Random Image)</button>
<button onclick="uploadDetect()">Upload Image</button>
<input type="file" id="fileInput" accept="image/*" style="display:none" onchange="handleUpload()">
</div>

<div class="section">
<h2>Benchmark & Optimize</h2>
<button onclick="runBenchmark()">Benchmark (50 iterations)</button>
<select id="quant"><option value="fp16">FP16</option><option value="int8">INT8</option><option value="fp32">FP32</option></select>
<button onclick="runOptimize()">Optimize Model</button>
</div>

<div id="output">Ready. Click "Detect" to run inference on a synthetic image.</div>
</div>

<script>
async function loadInfo(){
  try{const r=await fetch('/api/v1/model/info');const d=await r.json();
  document.getElementById('backend').textContent=d.backend;
  document.getElementById('classes').textContent=d.num_classes;
  document.getElementById('shape').textContent=d.input_shape[2]+'x'+d.input_shape[3];}catch(e){}
}

async function runDetection(){
  document.getElementById('output').textContent='Running inference...';
  try{const r=await fetch('/api/v1/detect/random?width=640&height=480',{method:'POST'});
  const d=await r.json();
  let out=`Inference Result (${d.backend})\\n${'='.repeat(50)}\\n`;
  out+=`Image: ${d.image_size[0]}x${d.image_size[1]}\\n`;
  out+=`Preprocessing: ${d.preprocessing_ms}ms\\nInference: ${d.inference_time_ms}ms\\nPostprocessing: ${d.postprocessing_ms}ms\\n`;
  out+=`Total: ${(d.preprocessing_ms+d.inference_time_ms+d.postprocessing_ms).toFixed(2)}ms\\n\\n`;
  out+=`Detections (${d.detections.length}):\\n`;
  d.detections.forEach((det,i)=>out+=`  ${i+1}. ${det.class_name} (${(det.confidence*100).toFixed(1)}%) bbox:[${det.bbox.map(b=>b.toFixed(3)).join(', ')}]\\n`);
  document.getElementById('output').textContent=out;}catch(e){document.getElementById('output').textContent='Error: '+e.message;}
}

function uploadDetect(){document.getElementById('fileInput').click()}
async function handleUpload(){
  const f=document.getElementById('fileInput').files[0];if(!f)return;
  const fd=new FormData();fd.append('file',f);
  document.getElementById('output').textContent='Processing image...';
  try{const r=await fetch('/api/v1/detect',{method:'POST',body:fd});
  const d=await r.json();
  let out=`Detection on: ${f.name}\\n${'='.repeat(50)}\\nInference: ${d.inference_time_ms}ms | Detections: ${d.detections.length}\\n\\n`;
  d.detections.forEach((det,i)=>out+=`  ${i+1}. ${det.class_name} (${(det.confidence*100).toFixed(1)}%)\\n`);
  document.getElementById('output').textContent=out;}catch(e){document.getElementById('output').textContent='Error: '+e.message;}
}

async function runBenchmark(){
  document.getElementById('output').textContent='Running benchmark (50 iterations)...';
  try{const r=await fetch('/api/v1/model/benchmark?iterations=50',{method:'POST'});
  const d=await r.json();
  let out=`Benchmark Results\\n${'='.repeat(50)}\\n`;
  out+=`Backend: ${d.backend}\\nIterations: ${d.iterations}\\n\\n`;
  out+=`Avg: ${d.avg_inference_ms}ms\\nMin: ${d.min_inference_ms}ms\\nMax: ${d.max_inference_ms}ms\\nP95: ${d.p95_inference_ms}ms\\n`;
  out+=`Throughput: ${d.throughput_fps} FPS\\nMemory: ${d.memory_usage_mb}MB\\n`;
  document.getElementById('output').textContent=out;}catch(e){document.getElementById('output').textContent='Error: '+e.message;}
}

async function runOptimize(){
  const q=document.getElementById('quant').value;
  document.getElementById('output').textContent='Optimizing...';
  try{const r=await fetch('/api/v1/model/optimize?quantization='+q,{method:'POST'});
  const d=await r.json();const o=d.optimization_result;
  let out=`Model Optimization\\n${'='.repeat(50)}\\n`;
  out+=`Original: ${o.original_size_mb}MB → Optimized: ${o.optimized_size_mb}MB (${o.size_reduction_pct}% smaller)\\n`;
  out+=`Estimated Speedup: ${o.estimated_speedup}\\n\\nOptimizations Applied:\\n`;
  o.optimizations_applied.forEach(op=>out+=`  - ${op}\\n`);
  out+=`\\nRecommendations:\\n`;d.recommendations.forEach(r=>out+=`  - ${r}\\n`);
  document.getElementById('output').textContent=out;}catch(e){document.getElementById('output').textContent='Error: '+e.message;}
}

loadInfo();
</script></body></html>"""


if __name__ == "__main__":
    import uvicorn
    uvicorn.run("main:app", host=config.HOST, port=config.PORT, reload=False)
